Rename dynalite base entity module
