+++ homeassistant/components/dynalite/entity.py
