+++ homeassistant/components/dynalite/switch.py
@@ -8,7 +8,7 @@ from homeassistant.const import STATE_ON
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .dynalitebase import DynaliteBase, async_setup_entry_base
+from .entity import DynaliteBase, async_setup_entry_base
 
 
 async def async_setup_entry(
